Attach a parent-metabolite mapping to a SDTM object

In case multiple analytes are measured for a specific administered drug, some functions need that information to correlate plasma concentrations with administrations.

Usage

add_metabolite_mapping(obj, pctestcd_parent, pctestcd_metabolite)

Arguments

obj

The SDTM object.

pctestcd_parent

The PCTESTCD of the parent compound.

pctestcd_metabolite

The PCTESTCD of the metabolite.

Value

The SDTM object.

See also

Examples

add_metabolite_mapping(examplinib_fe, "RS2023", "RS2023487A")
#> -------- SDTM data set summary -------- 
#> Study 2023000400 
#> 
#> Data disposition
#>   DOMAIN   SUBJECTS   OBSERVATIONS   
#>   dm       28         28             
#>   vs       28         56             
#>   ex       20         40             
#>   pc       20         1360           
#>   lb       28         28             
#>   pp       20         360            
#> 
#> Arms (DM):
#>   ACTARMCD   ACTARM           
#>   SCRNFAIL   Screen Failure   
#>   BA         Fed - Fasted     
#>   AB         Fasted - Fed     
#> 
#> Treatments (EX):
#>   EXAMPLINIB
#> 
#> PK sample specimens (PC):
#>   PLASMA
#> 
#> PK analytes (PC):
#>   PCTEST       PCTESTCD     
#>   RS2023       RS2023       
#>   RS2023487A   RS2023487A    
#>