Pharmaverse sample SDTM data

An sdtm object based on the CDISCPILOT01 data set exported from the 'pharmaversesdtm' package:

Usage

cdiscpilot01_sdtm

Format

cdiscpilot01_sdtm

A sdtm object with the following domains

dm

Demographics

ex

Exposure

vs

Vital signs

pc

Pharmacokinetic concentrations

pp

Pharmacokinetic parameters

lb

Laboratory data

ts

Trial summary

eg

ECG data

ae

Adverse events

Source

https://github.com/pharmaverse/pharmaversesdtma

Details

Patil L, Bundfuss S, Dahnert K, Gautier F, Mancini E, Namai T, Nguyen V, Shuliar V (2025). pharmaversesdtm: SDTM Test Data for the 'Pharmaverse'Family of Packages. R package version 1.3.1.9011, commit 55eba18f3d1d4b19e4fda886ff2da78582cae933, https://github.com/pharmaverse/pharmaversesdtm.

This is a fictional study with the title: "Safety and Efficacy of the Xanomeline Transdermal Therapeutic System (TTS) in Patients with Mild to Moderate Alzheimers Disease."