Synthesize a fictional EX domain for single dose administration

Synthesize a fictional EX domain for single dose administration

Usage

make_md_ex(
  dm,
  drug = "RS2023",
  dose = 500,
  treatment_duration = 50,
  missed_prob = 0.15,
  missed_doses = T,
  red_prob = 0.3
)

Arguments

dm

The DM including the subject info for whom to synthesize EX.

drug

The name of the drug to be administered.

dose

The dose.

treatment_duration

The treatment duration in days.

missed_prob

Probability to miss doses.

missed_doses

Switch whether to include randomly missed doses as boolean.

red_prob

The dose reduction probability.

Value

The EX domain as data frame.