Make a NIF data set from SDTM-formatted data

This function makes a basic NONMEM input file (NIF) data set from SDTM-formatted clinical study data following the conventions summarized in Bauer, CPT Pharmacometrics Syst. Pharmacol. (2019). For a more in-depth tutorial, see vignette("nif-vignette").

Usage

make_nif(
  sdtm_data,
  spec = NULL,
  impute_administration_time = TRUE,
  silent = FALSE,
  truncate_to_last_observation = TRUE,
  truncate_to_last_individual_obs = TRUE,
  use_pctptnum = TRUE
)

Arguments

sdtm_data

A sdtm object, i.e., essentially a list of SDTM domains as data tables. Typically, the SDTM data are loaded using read_sdtm_sas() or read_sdtm_xpt(). As a minimum, the following SDTM domains are needed: DM, VS, PC and EX.

spec

The sample specimen for the PC data as string (e.g., "BLOOD", "PLASMA", "URINE", "FECES"). When spec is NULL (default), the most likely specimen is selected.

impute_administration_time

A boolean value to indicate whether the time of administration is to be imputed from the PCRFDTC field from the PC domain. This field is 'permissible' and may not be present in certain SDTM data. The default is 'TRUE'.

silent

Boolean value to indicate whether warnings should be printed. The default is 'FALSE'.

truncate_to_last_observation

Boolean to indicate whether the data set should be truncated to the last observation. In this case, administrations after the last observation time point will deleted. The default is 'TRUE'.

truncate_to_last_individual_obs

Boolean to indicate whether observations should be truncted to the last individual observation.

use_pctptnum

Boolean to indicate whether to derive nominal time ('NTIME') from 'PCTPTNUM'.

Value

A NIF object.

Imputations

Subjects with administration but no observations for the respective analyte are deleted from the data set. For further imputations, see vignette("nif-imputations").

Output fields

ID Subject identification number

TIME Recorded time of administration or observation events in hours relative to the first individual event.

AMT Dose administered for dosing record, or zero for observations.

DOSE Dose in mg for administrations and post-dose observations.

DV The dependent variable, i.e., observed concentration, or zero for administration records, in mg/l.

LNDV The natural Log of DV.

RATE Rate of infusion of drug or zero if drug is given as a bolus.

MDV One for missing DV, else zero.

EVID Event ID: 0 for observations, 1 for administrations.

CMT Pharmacokinetic compartment. Will be set to 1 for administrations and 2 for observations. Should be changed afterwards, if needed.

DTC The date-time of the data record.

FIRSTDTC Date and time of first event per subject. This field is used internally for the calculation of TIME. Although it is not needed for NONMEM analysis, it is provided for subsequent NIF file building steps, e.g., addition of further time-dependent endpoints.

FIRSTADMINDTC The date-time of the first administration of the respective parent drug for the respective subject.

FIRSTTRTDTC The date-time of the first administration of any parent drug for the respective subject.

ANALYTE The analyte or drug in the data record.

TRTDY The treatment day, i.e., the relative day after the first treatment for the respective subject.

See also

Examples

make_nif(examplinib_fe)
#> No specimen specified. Set to PLASMA as the most likely.
#> Data cut-off was set to last observation time, 2001-03-12 10:13:00
#> NONMEM input file (NIF) data set with data from 1 studies
#> 680 observations from 20 subjects
#> Males: 14, females: 6 (30%)
#> 
#> Columns:
#> REF, ID, PERIOD, PCTESTCD, PCSTRESN, PCTEST, USUBJID, STUDYID, PCELTM, PCTPTNUM, PCTPT, PCDTC, PCSEQ, PCSPEC, PCRFTDTC, EPOCH, DTC, start.date, start.time, NTIME, EVID, CMT, AMT, DV, LNDV, MDV, ANALYTE, RATE, EXTRT, PARENT, METABOLITE, last_obs, EXDY, EXSTDTC, EXSTDY, EXENDY, EXENDTC, EXDOSE, EXROUTE, EXDOSFRM, EXSEQ, EXSTDTC_has_time, EXENDTC_has_time, end.date, end.time, date, time, DOSE, TYPE, dtc.date, SITEID, SUBJID, ACTARM, ACTARMCD, RFICDTC, RFSTDTC, SEX, AGE, AGEU, COUNTRY, ARM, ARMCD, RACE, ETHNIC, RFENDTC, HEIGHT, WEIGHT, BMI, FIRSTADMINDTC, FIRSTTRTDTC, FIRSTDTC, TIME, TRTDY, TAD 
#> 
#> First rows of NIF data (selected columns):
#> ID   NTIME   EVID   CMT   AMT   DV                    ANALYTE   DOSE   TIME   TAD   
#> 1    0       1      1     500   NA                    RS2023    500    0      0     
#> 1    0       0      2     0     0                     RS2023    NA     0      0     
#> 1    0.5     0      2     0     0.902503560799694     RS2023    500    0.5    0.5   
#> 1    1       0      2     0     1.6201458423574       RS2023    500    1      1     
#> 1    1.5     0      2     0     3.45719508560444      RS2023    500    1.5    1.5   
#> 1    2       0      2     0     3.02872200502414      RS2023    500    2      2     
#> 1    3       0      2     0     2.73796041824882      RS2023    500    3      3     
#> 1    4       0      2     0     1.57858378748303      RS2023    500    4      4     
#> 1    6       0      2     0     0.941951096527271     RS2023    500    6      6     
#> 1    8       0      2     0     0.459242625586454     RS2023    500    8      8     
#> 1    10      0      2     0     0.238956661290418     RS2023    500    10     10    
#> 1    12      0      2     0     0.201064807909193     RS2023    500    12     12    
#> 1    24      0      2     0     0.0364867151444544    RS2023    500    24     24    
#> 1    48      0      2     0     0.014237431673674     RS2023    500    48     48    
#> 1    72      0      2     0     0.00618959374465484   RS2023    500    72     72    
make_nif(examplinib_poc)
#> No specimen specified. Set to PLASMA as the most likely.
#> Data cut-off was set to last observation time, 2001-04-27 14:31:00
#> NONMEM input file (NIF) data set with data from 1 studies
#> 1344 observations from 80 subjects
#> Males: 51, females: 29 (36.2%)
#> 
#> Columns:
#> REF, ID, PCTESTCD, PCSTRESN, PCTEST, USUBJID, RFSTDTC.x, STUDYID, PCDTC, PCSPEC, EPOCH, PCTPT, PCTPTNUM, PCRFTDTC, PCSEQ, DTC, start.date, start.time, NTIME, EVID, CMT, AMT, DV, LNDV, MDV, ANALYTE, RATE, EXTRT, PARENT, METABOLITE, last_obs, EXSTDTC, EXENDTC, DOSE, EXDOSE, EXROUTE, EXDOSFRM, EXSEQ, EXSTDTC_has_time, EXENDTC_has_time, end.date, end.time, EXSTDY, EXENDY, date, time, EXDY, TYPE, dtc.date, SITEID, SUBJID, ACTARM, ACTARMCD, RFICDTC, RFSTDTC.y, SEX, AGE, AGEU, COUNTRY, ARM, ARMCD, RACE, ETHNIC, RFENDTC, HEIGHT, WEIGHT, BMI, FIRSTADMINDTC, FIRSTTRTDTC, FIRSTDTC, TIME, TRTDY, TAD 
#> 
#> First rows of NIF data (selected columns):
#> ID   NTIME   EVID   CMT   AMT   DV                   ANALYTE      DOSE   TIME    TAD    
#> 1    0       1      1     500   NA                   RS2023       500    0       0      
#> 1    0       0      2     0     0                    RS2023       NA     0       0      
#> 1    0       0      2     0     0                    RS2023487A   NA     0       0      
#> 1    1.5     0      2     0     3.36772568843661     RS2023       500    1.75    1.75   
#> 1    1.5     0      2     0     3.86720626722762     RS2023487A   500    1.75    1.75   
#> 1    4       0      2     0     0.973119920231158    RS2023       500    4.4     4.4    
#> 1    4       0      2     0     9.56502537142429     RS2023487A   500    4.4     4.4    
#> 1    0       1      1     500   NA                   RS2023       500    24      0      
#> 1    0       1      1     500   NA                   RS2023       500    48      0      
#> 1    0       1      1     500   NA                   RS2023       500    72      0      
#> 1    0       1      1     500   NA                   RS2023       500    96      0      
#> 1    0       1      1     500   NA                   RS2023       500    120     0      
#> 1    0       1      1     500   NA                   RS2023       500    144     0      
#> 1    0       1      1     500   NA                   RS2023       500    168     0      
#> 1    0       0      2     0     0.0362596629478831   RS2023       500    191.3   23.3