Make nominal time

Return NTIME lookup table or NULL if the xxELTM field is not included in the input data frame.

Usage

make_ntime(obj, include_day = FALSE, silent = NULL)

Arguments

obj

The input as data table.

include_day

include_day Include time component of treatment day, as logical.

silent

Suppress messages, as logical. Defaults to nif_option setting if NULL.

Value

A data frame.

Examples

make_ntime(examplinib_sad_nif)
#>    PCELTM NTIME
#> 1    <NA>    NA
#> 2    PT0H   0.0
#> 3  PT0.5H   0.5
#> 4    PT1H   1.0
#> 5  PT1.5H   1.5
#> 6    PT2H   2.0
#> 7    PT3H   3.0
#> 8    PT4H   4.0
#> 9    PT6H   6.0
#> 10   PT8H   8.0
#> 11  PT10H  10.0
#> 12  PT12H  12.0
#> 13  PT24H  24.0
#> 14  PT48H  48.0
#> 15  PT72H  72.0
#> 16  PT96H  96.0
#> 17 PT144H 144.0
#> 18 PT168H 168.0
make_ntime(examplinib_fe_nif)
#>    PCELTM NTIME
#> 1    <NA>    NA
#> 2    PT0H   0.0
#> 3  PT0.5H   0.5
#> 4    PT1H   1.0
#> 5  PT1.5H   1.5
#> 6    PT2H   2.0
#> 7    PT3H   3.0
#> 8    PT4H   4.0
#> 9    PT6H   6.0
#> 10   PT8H   8.0
#> 11  PT10H  10.0
#> 12  PT12H  12.0
#> 13  PT24H  24.0
#> 14  PT48H  48.0
#> 15  PT72H  72.0
#> 16  PT96H  96.0
#> 17 PT144H 144.0
#> 18 PT168H 168.0