Non-compartmental analysis

Non-compartmental analysis

Usage

nca(
  obj,
  analyte = NULL,
  parent = NULL,
  keep = "DOSE",
  group = NULL,
  nominal_time = FALSE,
  silent = FALSE,
  average_duplicates = TRUE
)

Arguments

obj

The source NIF object.

analyte

The analyte. If none specified and multiple analytes are in the dataset, defaults to the first analyte.

parent

The parent compound to derive the administration information from. By default, equal to the analyte.

keep

A vector of fields to keep in the output.

group

The grouping variable as string.

nominal_time

A boolean to indicate whether nominal time rather than actual time should be used for NCA.

silent

No message output.

average_duplicates

Boolean to indicate whether duplicate entries should be averaged.

Value

A data frame.

Examples

nca(examplinib_fe_nif)
#> NCA: No analyte specified. Selected RS2023 as the most likely.
#>    ID start end PPTESTCD   PPORRES exclude DOSE
#> 1   1     0 312  auclast 34.619531    <NA>  500
#> 2   1     0 312     cmax  4.117052    <NA>  500
#> 3   1     0 312     tmax  2.000000    <NA>  500
#> 4   2     0 312  auclast 21.021256    <NA>  500
#> 5   2     0 312     cmax  2.584639    <NA>  500
#> 6   2     0 312     tmax  3.000000    <NA>  500
#> 7   3     0 312  auclast 26.098432    <NA>  500
#> 8   3     0 312     cmax  3.521993    <NA>  500
#> 9   3     0 312     tmax  3.000000    <NA>  500
#> 10  4     0 312  auclast 25.647595    <NA>  500
#> 11  4     0 312     cmax  3.093239    <NA>  500
#> 12  4     0 312     tmax  3.000000    <NA>  500
#> 13  5     0 312  auclast 23.025916    <NA>  500
#> 14  5     0 312     cmax  2.353615    <NA>  500
#> 15  5     0 312     tmax  4.000000    <NA>  500
#> 16  6     0 312  auclast 26.465487    <NA>  500
#> 17  6     0 312     cmax  4.301667    <NA>  500
#> 18  6     0 312     tmax  2.000000    <NA>  500
#> 19  7     0 312  auclast 21.709273    <NA>  500
#> 20  7     0 312     cmax  2.544173    <NA>  500
#> 21  7     0 312     tmax  3.000000    <NA>  500
#> 22  8     0 312  auclast 26.925087    <NA>  500
#> 23  8     0 312     cmax  3.827967    <NA>  500
#> 24  8     0 312     tmax  2.000000    <NA>  500
#> 25  9     0 312  auclast 20.368540    <NA>  500
#> 26  9     0 312     cmax  3.309859    <NA>  500
#> 27  9     0 312     tmax  3.000000    <NA>  500
#> 28 10     0 312  auclast 14.818389    <NA>  500
#> 29 10     0 312     cmax  3.684004    <NA>  500
#> 30 10     0 312     tmax  1.500000    <NA>  500
#> 31 11     0 312  auclast 17.072351    <NA>  500
#> 32 11     0 312     cmax  2.969615    <NA>  500
#> 33 11     0 312     tmax  3.000000    <NA>  500
#> 34 12     0 312  auclast 31.462430    <NA>  500
#> 35 12     0 312     cmax  3.368056    <NA>  500
#> 36 12     0 312     tmax  4.000000    <NA>  500
#> 37 13     0 312  auclast 25.008302    <NA>  500
#> 38 13     0 312     cmax  3.913508    <NA>  500
#> 39 13     0 312     tmax  2.000000    <NA>  500
#> 40 14     0 312  auclast 25.840839    <NA>  500
#> 41 14     0 312     cmax  3.356072    <NA>  500
#> 42 14     0 312     tmax  3.000000    <NA>  500
#> 43 15     0 312  auclast 33.373488    <NA>  500
#> 44 15     0 312     cmax  2.495331    <NA>  500
#> 45 15     0 312     tmax  4.000000    <NA>  500
#> 46 16     0 312  auclast 30.947423    <NA>  500
#> 47 16     0 312     cmax  4.136596    <NA>  500
#> 48 16     0 312     tmax  3.000000    <NA>  500
#> 49 17     0 312  auclast 41.999082    <NA>  500
#> 50 17     0 312     cmax  2.896972    <NA>  500
#> 51 17     0 312     tmax  3.000000    <NA>  500
#> 52 18     0 312  auclast 32.411727    <NA>  500
#> 53 18     0 312     cmax  3.490646    <NA>  500
#> 54 18     0 312     tmax  3.000000    <NA>  500
#> 55 19     0 312  auclast 23.380969    <NA>  500
#> 56 19     0 312     cmax  3.904765    <NA>  500
#> 57 19     0 312     tmax  2.000000    <NA>  500
#> 58 20     0 312  auclast 55.335799    <NA>  500
#> 59 20     0 312     cmax  4.909493    <NA>  500
#> 60 20     0 312     tmax  2.000000    <NA>  500
nca(examplinib_fe_nif, group = c("FASTED", "SEX"), analyte = "RS2023")
#> NCA: Group by FASTED+SEX
#>     FASTED SEX ID start end            PPTESTCD      PPORRES exclude DOSE
#> 1        0   0  1     0  24             auclast 2.844165e+01    <NA>  500
#> 2        0   0  1     0 Inf                cmax 4.117052e+00    <NA>  500
#> 3        0   0  1     0 Inf                tmax 2.000000e+00    <NA>  500
#> 4        0   0  1     0 Inf               tlast 1.680000e+02    <NA>  500
#> 5        0   0  1     0 Inf           clast.obs 2.306155e-03    <NA>  500
#> 6        0   0  1     0 Inf            lambda.z 3.008553e-02    <NA>  500
#> 7        0   0  1     0 Inf           r.squared 9.946787e-01    <NA>  500
#> 8        0   0  1     0 Inf       adj.r.squared 9.929049e-01    <NA>  500
#> 9        0   0  1     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 10       0   0  1     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 11       0   0  1     0 Inf          clast.pred 2.115275e-03    <NA>  500
#> 12       0   0  1     0 Inf           half.life 2.303922e+01    <NA>  500
#> 13       0   0  1     0 Inf          span.ratio 5.208509e+00    <NA>  500
#> 14       0   0  1     0 Inf          aucinf.obs 3.461788e+01    <NA>  500
#> 15       1   0  1   312 336             auclast 2.879043e+01    <NA>  500
#> 16       1   0  1   312 Inf                cmax 2.786793e+00    <NA>  500
#> 17       1   0  1   312 Inf                tmax 4.000000e+00    <NA>  500
#> 18       1   0  1   312 Inf               tlast 1.680000e+02    <NA>  500
#> 19       1   0  1   312 Inf           clast.obs 2.286446e-03    <NA>  500
#> 20       1   0  1   312 Inf            lambda.z 3.119554e-02    <NA>  500
#> 21       1   0  1   312 Inf           r.squared 9.974720e-01    <NA>  500
#> 22       1   0  1   312 Inf       adj.r.squared 9.949440e-01    <NA>  500
#> 23       1   0  1   312 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 24       1   0  1   312 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 25       1   0  1   312 Inf          clast.pred 2.388154e-03    <NA>  500
#> 26       1   0  1   312 Inf           half.life 2.221943e+01    <NA>  500
#> 27       1   0  1   312 Inf          span.ratio 3.240407e+00    <NA>  500
#> 28       1   0  1   312 Inf          aucinf.obs 3.515430e+01    <NA>  500
#> 29       1   1  2     0  24             auclast 1.869313e+01    <NA>  500
#> 30       1   1  2     0 Inf                cmax 2.584639e+00    <NA>  500
#> 31       1   1  2     0 Inf                tmax 3.000000e+00    <NA>  500
#> 32       1   1  2     0 Inf               tlast 1.680000e+02    <NA>  500
#> 33       1   1  2     0 Inf           clast.obs 2.779978e-04    <NA>  500
#> 34       1   1  2     0 Inf            lambda.z 3.876095e-02    <NA>  500
#> 35       1   1  2     0 Inf           r.squared 9.989176e-01    <NA>  500
#> 36       1   1  2     0 Inf       adj.r.squared 9.985568e-01    <NA>  500
#> 37       1   1  2     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 38       1   1  2     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 39       1   1  2     0 Inf          clast.pred 2.866802e-04    <NA>  500
#> 40       1   1  2     0 Inf           half.life 1.788262e+01    <NA>  500
#> 41       1   1  2     0 Inf          span.ratio 6.710428e+00    <NA>  500
#> 42       1   1  2     0 Inf          aucinf.obs 2.102126e+01    <NA>  500
#> 43       0   1  2   312 336             auclast 1.876529e+01    <NA>  500
#> 44       0   1  2   312 Inf                cmax 4.207926e+00    <NA>  500
#> 45       0   1  2   312 Inf                tmax 2.000000e+00    <NA>  500
#> 46       0   1  2   312 Inf               tlast 1.680000e+02    <NA>  500
#> 47       0   1  2   312 Inf           clast.obs 2.887684e-04    <NA>  500
#> 48       0   1  2   312 Inf            lambda.z 3.846851e-02    <NA>  500
#> 49       0   1  2   312 Inf           r.squared 9.999475e-01    <NA>  500
#> 50       0   1  2   312 Inf       adj.r.squared 9.999300e-01    <NA>  500
#> 51       0   1  2   312 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 52       0   1  2   312 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 53       0   1  2   312 Inf          clast.pred 2.898377e-04    <NA>  500
#> 54       0   1  2   312 Inf           half.life 1.801856e+01    <NA>  500
#> 55       0   1  2   312 Inf          span.ratio 6.659799e+00    <NA>  500
#> 56       0   1  2   312 Inf          aucinf.obs 2.079839e+01    <NA>  500
#> 57       1   0  3     0  24             auclast 2.360760e+01    <NA>  500
#> 58       1   0  3     0 Inf                cmax 3.521993e+00    <NA>  500
#> 59       1   0  3     0 Inf                tmax 3.000000e+00    <NA>  500
#> 60       1   0  3     0 Inf               tlast 1.680000e+02    <NA>  500
#> 61       1   0  3     0 Inf           clast.obs 9.773726e-04    <NA>  500
#> 62       1   0  3     0 Inf            lambda.z 2.895954e-02    <NA>  500
#> 63       1   0  3     0 Inf           r.squared 9.987673e-01    <NA>  500
#> 64       1   0  3     0 Inf       adj.r.squared 9.983564e-01    <NA>  500
#> 65       1   0  3     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 66       1   0  3     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 67       1   0  3     0 Inf          clast.pred 1.038087e-03    <NA>  500
#> 68       1   0  3     0 Inf           half.life 2.393502e+01    <NA>  500
#> 69       1   0  3     0 Inf          span.ratio 5.013574e+00    <NA>  500
#> 70       1   0  3     0 Inf          aucinf.obs 2.609489e+01    <NA>  500
#> 71       0   0  3   312 336             auclast 2.169573e+01    <NA>  500
#> 72       0   0  3   312 Inf                cmax 4.017258e+00    <NA>  500
#> 73       0   0  3   312 Inf                tmax 2.000000e+00    <NA>  500
#> 74       0   0  3   312 Inf               tlast 1.680000e+02    <NA>  500
#> 75       0   0  3   312 Inf           clast.obs 1.223981e-03    <NA>  500
#> 76       0   0  3   312 Inf            lambda.z 2.695583e-02    <NA>  500
#> 77       0   0  3   312 Inf           r.squared 9.994732e-01    <NA>  500
#> 78       0   0  3   312 Inf       adj.r.squared 9.992976e-01    <NA>  500
#> 79       0   0  3   312 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 80       0   0  3   312 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 81       0   0  3   312 Inf          clast.pred 1.271309e-03    <NA>  500
#> 82       0   0  3   312 Inf           half.life 2.571418e+01    <NA>  500
#> 83       0   0  3   312 Inf          span.ratio 4.666685e+00    <NA>  500
#> 84       0   0  3   312 Inf          aucinf.obs 2.421216e+01    <NA>  500
#> 85       0   1  4     0  24             auclast 2.225683e+01    <NA>  500
#> 86       0   1  4     0 Inf                cmax 3.093239e+00    <NA>  500
#> 87       0   1  4     0 Inf                tmax 3.000000e+00    <NA>  500
#> 88       0   1  4     0 Inf               tlast 1.680000e+02    <NA>  500
#> 89       0   1  4     0 Inf           clast.obs 1.831305e-04    <NA>  500
#> 90       0   1  4     0 Inf            lambda.z 4.759060e-02    <NA>  500
#> 91       0   1  4     0 Inf           r.squared 9.995322e-01    <NA>  500
#> 92       0   1  4     0 Inf       adj.r.squared 9.994152e-01    <NA>  500
#> 93       0   1  4     0 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 94       0   1  4     0 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 95       0   1  4     0 Inf          clast.pred 1.698768e-04    <NA>  500
#> 96       0   1  4     0 Inf           half.life 1.456479e+01    <NA>  500
#> 97       0   1  4     0 Inf          span.ratio 9.886856e+00    <NA>  500
#> 98       0   1  4     0 Inf          aucinf.obs 2.564772e+01    <NA>  500
#> 99       1   1  4   312 336             auclast 2.379202e+01    <NA>  500
#> 100      1   1  4   312 Inf                cmax 3.220627e+00    <NA>  500
#> 101      1   1  4   312 Inf                tmax 4.000000e+00    <NA>  500
#> 102      1   1  4   312 Inf               tlast 1.680000e+02    <NA>  500
#> 103      1   1  4   312 Inf           clast.obs 1.579296e-04    <NA>  500
#> 104      1   1  4   312 Inf            lambda.z 4.825993e-02    <NA>  500
#> 105      1   1  4   312 Inf           r.squared 9.993005e-01    <NA>  500
#> 106      1   1  4   312 Inf       adj.r.squared 9.989507e-01    <NA>  500
#> 107      1   1  4   312 Inf lambda.z.time.first 7.200000e+01    <NA>  500
#> 108      1   1  4   312 Inf   lambda.z.n.points 4.000000e+00    <NA>  500
#> 109      1   1  4   312 Inf          clast.pred 1.662031e-04    <NA>  500
#> 110      1   1  4   312 Inf           half.life 1.436279e+01    <NA>  500
#> 111      1   1  4   312 Inf          span.ratio 6.683938e+00    <NA>  500
#> 112      1   1  4   312 Inf          aucinf.obs 2.719199e+01    <NA>  500
#> 113      1   1  5     0  24             auclast 2.052055e+01    <NA>  500
#> 114      1   1  5     0 Inf                cmax 2.353615e+00    <NA>  500
#> 115      1   1  5     0 Inf                tmax 4.000000e+00    <NA>  500
#> 116      1   1  5     0 Inf               tlast 1.680000e+02    <NA>  500
#> 117      1   1  5     0 Inf           clast.obs 3.876432e-04    <NA>  500
#> 118      1   1  5     0 Inf            lambda.z 3.727848e-02    <NA>  500
#> 119      1   1  5     0 Inf           r.squared 9.990067e-01    <NA>  500
#> 120      1   1  5     0 Inf       adj.r.squared 9.986756e-01    <NA>  500
#> 121      1   1  5     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 122      1   1  5     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 123      1   1  5     0 Inf          clast.pred 3.741296e-04    <NA>  500
#> 124      1   1  5     0 Inf           half.life 1.859376e+01    <NA>  500
#> 125      1   1  5     0 Inf          span.ratio 6.453777e+00    <NA>  500
#> 126      1   1  5     0 Inf          aucinf.obs 2.302619e+01    <NA>  500
#> 127      0   1  5   312 336             auclast 1.917215e+01    <NA>  500
#> 128      0   1  5   312 Inf                cmax 3.419486e+00    <NA>  500
#> 129      0   1  5   312 Inf                tmax 3.000000e+00    <NA>  500
#> 130      0   1  5   312 Inf               tlast 1.680000e+02    <NA>  500
#> 131      0   1  5   312 Inf           clast.obs 3.262676e-04    <NA>  500
#> 132      0   1  5   312 Inf            lambda.z 3.892436e-02    <NA>  500
#> 133      0   1  5   312 Inf           r.squared 9.995921e-01    <NA>  500
#> 134      0   1  5   312 Inf       adj.r.squared 9.994561e-01    <NA>  500
#> 135      0   1  5   312 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 136      0   1  5   312 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 137      0   1  5   312 Inf          clast.pred 3.415926e-04    <NA>  500
#> 138      0   1  5   312 Inf           half.life 1.780754e+01    <NA>  500
#> 139      0   1  5   312 Inf          span.ratio 6.738718e+00    <NA>  500
#> 140      0   1  5   312 Inf          aucinf.obs 2.140998e+01    <NA>  500
#> 141      0   1  6     0  24             auclast 2.338414e+01    <NA>  500
#> 142      0   1  6     0 Inf                cmax 4.301667e+00    <NA>  500
#> 143      0   1  6     0 Inf                tmax 2.000000e+00    <NA>  500
#> 144      0   1  6     0 Inf               tlast 1.680000e+02    <NA>  500
#> 145      0   1  6     0 Inf           clast.obs 4.183156e-04    <NA>  500
#> 146      0   1  6     0 Inf            lambda.z 3.910884e-02    <NA>  500
#> 147      0   1  6     0 Inf           r.squared 9.995321e-01    <NA>  500
#> 148      0   1  6     0 Inf       adj.r.squared 9.993761e-01    <NA>  500
#> 149      0   1  6     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 150      0   1  6     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 151      0   1  6     0 Inf          clast.pred 4.066737e-04    <NA>  500
#> 152      0   1  6     0 Inf           half.life 1.772354e+01    <NA>  500
#> 153      0   1  6     0 Inf          span.ratio 6.770656e+00    <NA>  500
#> 154      0   1  6     0 Inf          aucinf.obs 2.646519e+01    <NA>  500
#> 155      1   1  6   312 336             auclast 2.226588e+01    <NA>  500
#> 156      1   1  6   312 Inf                cmax 2.315952e+00    <NA>  500
#> 157      1   1  6   312 Inf                tmax 4.000000e+00    <NA>  500
#> 158      1   1  6   312 Inf               tlast 1.680000e+02    <NA>  500
#> 159      1   1  6   312 Inf           clast.obs 4.982326e-04    <NA>  500
#> 160      1   1  6   312 Inf            lambda.z 3.644749e-02    <NA>  500
#> 161      1   1  6   312 Inf           r.squared 9.999479e-01    <NA>  500
#> 162      1   1  6   312 Inf       adj.r.squared 9.998959e-01    <NA>  500
#> 163      1   1  6   312 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 164      1   1  6   312 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 165      1   1  6   312 Inf          clast.pred 4.946147e-04    <NA>  500
#> 166      1   1  6   312 Inf           half.life 1.901769e+01    <NA>  500
#> 167      1   1  6   312 Inf          span.ratio 3.785948e+00    <NA>  500
#> 168      1   1  6   312 Inf          aucinf.obs 2.618474e+01    <NA>  500
#> 169      1   0  7     0  24             auclast 1.927248e+01    <NA>  500
#> 170      1   0  7     0 Inf                cmax 2.544173e+00    <NA>  500
#> 171      1   0  7     0 Inf                tmax 3.000000e+00    <NA>  500
#> 172      1   0  7     0 Inf               tlast 1.680000e+02    <NA>  500
#> 173      1   0  7     0 Inf           clast.obs 3.302869e-04    <NA>  500
#> 174      1   0  7     0 Inf            lambda.z 3.852151e-02    <NA>  500
#> 175      1   0  7     0 Inf           r.squared 9.991487e-01    <NA>  500
#> 176      1   0  7     0 Inf       adj.r.squared 9.988650e-01    <NA>  500
#> 177      1   0  7     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 178      1   0  7     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 179      1   0  7     0 Inf          clast.pred 3.369864e-04    <NA>  500
#> 180      1   0  7     0 Inf           half.life 1.799377e+01    <NA>  500
#> 181      1   0  7     0 Inf          span.ratio 6.668975e+00    <NA>  500
#> 182      1   0  7     0 Inf          aucinf.obs 2.170840e+01    <NA>  500
#> 183      0   0  7   312 336             auclast 2.099345e+01    <NA>  500
#> 184      0   0  7   312 Inf                cmax 3.845942e+00    <NA>  500
#> 185      0   0  7   312 Inf                tmax 2.000000e+00    <NA>  500
#> 186      0   0  7   312 Inf               tlast 1.680000e+02    <NA>  500
#> 187      0   0  7   312 Inf           clast.obs 3.032569e-04    <NA>  500
#> 188      0   0  7   312 Inf            lambda.z 3.824063e-02    <NA>  500
#> 189      0   0  7   312 Inf           r.squared 9.969460e-01    <NA>  500
#> 190      0   0  7   312 Inf       adj.r.squared 9.961826e-01    <NA>  500
#> 191      0   0  7   312 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 192      0   0  7   312 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 193      0   0  7   312 Inf          clast.pred 3.507520e-04    <NA>  500
#> 194      0   0  7   312 Inf           half.life 1.812594e+01    <NA>  500
#> 195      0   0  7   312 Inf          span.ratio 7.944417e+00    <NA>  500
#> 196      0   0  7   312 Inf          aucinf.obs 2.322382e+01    <NA>  500
#> 197      0   0  8     0  24             auclast 2.340205e+01    <NA>  500
#> 198      0   0  8     0 Inf                cmax 3.827967e+00    <NA>  500
#> 199      0   0  8     0 Inf                tmax 2.000000e+00    <NA>  500
#> 200      0   0  8     0 Inf               tlast 1.680000e+02    <NA>  500
#> 201      0   0  8     0 Inf           clast.obs 1.221369e-03    <NA>  500
#> 202      0   0  8     0 Inf            lambda.z 3.096207e-02    <NA>  500
#> 203      0   0  8     0 Inf           r.squared 9.975363e-01    <NA>  500
#> 204      0   0  8     0 Inf       adj.r.squared 9.967150e-01    <NA>  500
#> 205      0   0  8     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 206      0   0  8     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 207      0   0  8     0 Inf          clast.pred 1.139104e-03    <NA>  500
#> 208      0   0  8     0 Inf           half.life 2.238698e+01    <NA>  500
#> 209      0   0  8     0 Inf          span.ratio 5.360259e+00    <NA>  500
#> 210      0   0  8     0 Inf          aucinf.obs 2.692676e+01    <NA>  500
#> 211      1   0  8   312 336             auclast 2.384263e+01    <NA>  500
#> 212      1   0  8   312 Inf                cmax 3.094755e+00    <NA>  500
#> 213      1   0  8   312 Inf                tmax 2.000000e+00    <NA>  500
#> 214      1   0  8   312 Inf               tlast 1.680000e+02    <NA>  500
#> 215      1   0  8   312 Inf           clast.obs 1.186172e-03    <NA>  500
#> 216      1   0  8   312 Inf            lambda.z 3.360508e-02    <NA>  500
#> 217      1   0  8   312 Inf           r.squared 9.999764e-01    <NA>  500
#> 218      1   0  8   312 Inf       adj.r.squared 9.999528e-01    <NA>  500
#> 219      1   0  8   312 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 220      1   0  8   312 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 221      1   0  8   312 Inf          clast.pred 1.191551e-03    <NA>  500
#> 222      1   0  8   312 Inf           half.life 2.062626e+01    <NA>  500
#> 223      1   0  8   312 Inf          span.ratio 3.490696e+00    <NA>  500
#> 224      1   0  8   312 Inf          aucinf.obs 2.726259e+01    <NA>  500
#> 225      0   1  9     0  24             auclast 1.838849e+01    <NA>  500
#> 226      0   1  9     0 Inf                cmax 3.309859e+00    <NA>  500
#> 227      0   1  9     0 Inf                tmax 3.000000e+00    <NA>  500
#> 228      0   1  9     0 Inf               tlast 1.680000e+02    <NA>  500
#> 229      0   1  9     0 Inf           clast.obs 5.882321e-05    <NA>  500
#> 230      0   1  9     0 Inf            lambda.z 5.228017e-02    <NA>  500
#> 231      0   1  9     0 Inf           r.squared 9.997683e-01    <NA>  500
#> 232      0   1  9     0 Inf       adj.r.squared 9.997103e-01    <NA>  500
#> 233      0   1  9     0 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 234      0   1  9     0 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 235      0   1  9     0 Inf          clast.pred 5.517922e-05    <NA>  500
#> 236      0   1  9     0 Inf           half.life 1.325832e+01    <NA>  500
#> 237      0   1  9     0 Inf          span.ratio 1.086111e+01    <NA>  500
#> 238      0   1  9     0 Inf          aucinf.obs 2.036856e+01    <NA>  500
#> 239      1   1  9   312 336             auclast 1.799416e+01    <NA>  500
#> 240      1   1  9   312 Inf                cmax 2.351921e+00    <NA>  500
#> 241      1   1  9   312 Inf                tmax 4.000000e+00    <NA>  500
#> 242      1   1  9   312 Inf               tlast 1.680000e+02    <NA>  500
#> 243      1   1  9   312 Inf           clast.obs 7.396393e-05    <NA>  500
#> 244      1   1  9   312 Inf            lambda.z 4.910548e-02    <NA>  500
#> 245      1   1  9   312 Inf           r.squared 9.999731e-01    <NA>  500
#> 246      1   1  9   312 Inf       adj.r.squared 9.999461e-01    <NA>  500
#> 247      1   1  9   312 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 248      1   1  9   312 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 249      1   1  9   312 Inf          clast.pred 7.344337e-05    <NA>  500
#> 250      1   1  9   312 Inf           half.life 1.411548e+01    <NA>  500
#> 251      1   1  9   312 Inf          span.ratio 5.100784e+00    <NA>  500
#> 252      1   1  9   312 Inf          aucinf.obs 2.010395e+01    <NA>  500
#> 253      0   1 10     0  24             auclast 1.395272e+01    <NA>  500
#> 254      0   1 10     0 Inf                cmax 3.684004e+00    <NA>  500
#> 255      0   1 10     0 Inf                tmax 1.500000e+00    <NA>  500
#> 256      0   1 10     0 Inf               tlast 1.680000e+02    <NA>  500
#> 257      0   1 10     0 Inf           clast.obs 2.298161e-04    <NA>  500
#> 258      0   1 10     0 Inf            lambda.z 3.312629e-02    <NA>  500
#> 259      0   1 10     0 Inf           r.squared 9.984700e-01    <NA>  500
#> 260      0   1 10     0 Inf       adj.r.squared 9.980875e-01    <NA>  500
#> 261      0   1 10     0 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 262      0   1 10     0 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 263      0   1 10     0 Inf          clast.pred 2.457976e-04    <NA>  500
#> 264      0   1 10     0 Inf           half.life 2.092438e+01    <NA>  500
#> 265      0   1 10     0 Inf          span.ratio 6.881923e+00    <NA>  500
#> 266      0   1 10     0 Inf          aucinf.obs 1.481809e+01    <NA>  500
#> 267      1   1 10   312 336             auclast 1.468681e+01    <NA>  500
#> 268      1   1 10   312 Inf                cmax 2.279072e+00    <NA>  500
#> 269      1   1 10   312 Inf                tmax 4.000000e+00    <NA>  500
#> 270      1   1 10   312 Inf               tlast 1.680000e+02    <NA>  500
#> 271      1   1 10   312 Inf           clast.obs 2.869013e-04    <NA>  500
#> 272      1   1 10   312 Inf            lambda.z 3.224868e-02    <NA>  500
#> 273      1   1 10   312 Inf           r.squared 9.992360e-01    <NA>  500
#> 274      1   1 10   312 Inf       adj.r.squared 9.989814e-01    <NA>  500
#> 275      1   1 10   312 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 276      1   1 10   312 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 277      1   1 10   312 Inf          clast.pred 2.713300e-04    <NA>  500
#> 278      1   1 10   312 Inf           half.life 2.149381e+01    <NA>  500
#> 279      1   1 10   312 Inf          span.ratio 5.583002e+00    <NA>  500
#> 280      1   1 10   312 Inf          aucinf.obs 1.564319e+01    <NA>  500
#> 281      1   1 11     0  24             auclast 1.585721e+01    <NA>  500
#> 282      1   1 11     0 Inf                cmax 2.969615e+00    <NA>  500
#> 283      1   1 11     0 Inf                tmax 3.000000e+00    <NA>  500
#> 284      1   1 11     0 Inf               tlast 1.680000e+02    <NA>  500
#> 285      1   1 11     0 Inf           clast.obs 2.034536e-04    <NA>  500
#> 286      1   1 11     0 Inf            lambda.z 3.522838e-02    <NA>  500
#> 287      1   1 11     0 Inf           r.squared 9.998696e-01    <NA>  500
#> 288      1   1 11     0 Inf       adj.r.squared 9.997392e-01    <NA>  500
#> 289      1   1 11     0 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 290      1   1 11     0 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 291      1   1 11     0 Inf          clast.pred 2.057345e-04    <NA>  500
#> 292      1   1 11     0 Inf           half.life 1.967582e+01    <NA>  500
#> 293      1   1 11     0 Inf          span.ratio 3.659315e+00    <NA>  500
#> 294      1   1 11     0 Inf          aucinf.obs 1.707292e+01    <NA>  500
#> 295      0   1 11   312 336             auclast 1.510834e+01    <NA>  500
#> 296      0   1 11   312 Inf                cmax 3.846661e+00    <NA>  500
#> 297      0   1 11   312 Inf                tmax 1.500000e+00    <NA>  500
#> 298      0   1 11   312 Inf               tlast 1.680000e+02    <NA>  500
#> 299      0   1 11   312 Inf           clast.obs 1.954281e-04    <NA>  500
#> 300      0   1 11   312 Inf            lambda.z 3.864208e-02    <NA>  500
#> 301      0   1 11   312 Inf           r.squared 9.980942e-01    <NA>  500
#> 302      0   1 11   312 Inf       adj.r.squared 9.976177e-01    <NA>  500
#> 303      0   1 11   312 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 304      0   1 11   312 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 305      0   1 11   312 Inf          clast.pred 1.769922e-04    <NA>  500
#> 306      0   1 11   312 Inf           half.life 1.793763e+01    <NA>  500
#> 307      0   1 11   312 Inf          span.ratio 8.027818e+00    <NA>  500
#> 308      0   1 11   312 Inf          aucinf.obs 1.630335e+01    <NA>  500
#> 309      1   1 12     0  24             auclast 2.629916e+01    <NA>  500
#> 310      1   1 12     0 Inf                cmax 3.368056e+00    <NA>  500
#> 311      1   1 12     0 Inf                tmax 4.000000e+00    <NA>  500
#> 312      1   1 12     0 Inf               tlast 1.680000e+02    <NA>  500
#> 313      1   1 12     0 Inf           clast.obs 7.042156e-04    <NA>  500
#> 314      1   1 12     0 Inf            lambda.z 3.928609e-02    <NA>  500
#> 315      1   1 12     0 Inf           r.squared 9.997578e-01    <NA>  500
#> 316      1   1 12     0 Inf       adj.r.squared 9.996972e-01    <NA>  500
#> 317      1   1 12     0 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 318      1   1 12     0 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 319      1   1 12     0 Inf          clast.pred 7.023768e-04    <NA>  500
#> 320      1   1 12     0 Inf           half.life 1.764358e+01    <NA>  500
#> 321      1   1 12     0 Inf          span.ratio 8.161611e+00    <NA>  500
#> 322      1   1 12     0 Inf          aucinf.obs 3.146301e+01    <NA>  500
#> 323      0   1 12   312 336             auclast 2.719137e+01    <NA>  500
#> 324      0   1 12   312 Inf                cmax 3.567580e+00    <NA>  500
#> 325      0   1 12   312 Inf                tmax 2.000000e+00    <NA>  500
#> 326      0   1 12   312 Inf               tlast 1.680000e+02    <NA>  500
#> 327      0   1 12   312 Inf           clast.obs 6.562333e-04    <NA>  500
#> 328      0   1 12   312 Inf            lambda.z 4.210543e-02    <NA>  500
#> 329      0   1 12   312 Inf           r.squared 9.999468e-01    <NA>  500
#> 330      0   1 12   312 Inf       adj.r.squared 9.998936e-01    <NA>  500
#> 331      0   1 12   312 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 332      0   1 12   312 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 333      0   1 12   312 Inf          clast.pred 6.506718e-04    <NA>  500
#> 334      0   1 12   312 Inf           half.life 1.646218e+01    <NA>  500
#> 335      0   1 12   312 Inf          span.ratio 4.373661e+00    <NA>  500
#> 336      0   1 12   312 Inf          aucinf.obs 3.326074e+01    <NA>  500
#> 337      0   1 13     0  24             auclast 2.229471e+01    <NA>  500
#> 338      0   1 13     0 Inf                cmax 3.913508e+00    <NA>  500
#> 339      0   1 13     0 Inf                tmax 2.000000e+00    <NA>  500
#> 340      0   1 13     0 Inf               tlast 1.680000e+02    <NA>  500
#> 341      0   1 13     0 Inf           clast.obs 1.232440e-03    <NA>  500
#> 342      0   1 13     0 Inf            lambda.z 2.948262e-02    <NA>  500
#> 343      0   1 13     0 Inf           r.squared 9.999892e-01    <NA>  500
#> 344      0   1 13     0 Inf       adj.r.squared 9.999785e-01    <NA>  500
#> 345      0   1 13     0 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 346      0   1 13     0 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 347      0   1 13     0 Inf          clast.pred 1.229141e-03    <NA>  500
#> 348      0   1 13     0 Inf           half.life 2.351037e+01    <NA>  500
#> 349      0   1 13     0 Inf          span.ratio 3.062479e+00    <NA>  500
#> 350      0   1 13     0 Inf          aucinf.obs 2.500641e+01    <NA>  500
#> 351      1   1 13   312 336             auclast 2.367578e+01    <NA>  500
#> 352      1   1 13   312 Inf                cmax 3.429618e+00    <NA>  500
#> 353      1   1 13   312 Inf                tmax 3.000000e+00    <NA>  500
#> 354      1   1 13   312 Inf               tlast 1.680000e+02    <NA>  500
#> 355      1   1 13   312 Inf           clast.obs 1.239233e-03    <NA>  500
#> 356      1   1 13   312 Inf            lambda.z 3.160307e-02    <NA>  500
#> 357      1   1 13   312 Inf           r.squared 9.926786e-01    <NA>  500
#> 358      1   1 13   312 Inf       adj.r.squared 9.908482e-01    <NA>  500
#> 359      1   1 13   312 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 360      1   1 13   312 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 361      1   1 13   312 Inf          clast.pred 9.879283e-04    <NA>  500
#> 362      1   1 13   312 Inf           half.life 2.193290e+01    <NA>  500
#> 363      1   1 13   312 Inf          span.ratio 6.565478e+00    <NA>  500
#> 364      1   1 13   312 Inf          aucinf.obs 2.667545e+01    <NA>  500
#> 365      1   1 14     0  24             auclast 2.298400e+01    <NA>  500
#> 366      1   1 14     0 Inf                cmax 3.356072e+00    <NA>  500
#> 367      1   1 14     0 Inf                tmax 3.000000e+00    <NA>  500
#> 368      1   1 14     0 Inf               tlast 1.680000e+02    <NA>  500
#> 369      1   1 14     0 Inf           clast.obs 1.236824e-03    <NA>  500
#> 370      1   1 14     0 Inf            lambda.z 2.709298e-02    <NA>  500
#> 371      1   1 14     0 Inf           r.squared 1.000000e+00    <NA>  500
#> 372      1   1 14     0 Inf       adj.r.squared 9.999999e-01    <NA>  500
#> 373      1   1 14     0 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 374      1   1 14     0 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 375      1   1 14     0 Inf          clast.pred 1.237004e-03    <NA>  500
#> 376      1   1 14     0 Inf           half.life 2.558402e+01    <NA>  500
#> 377      1   1 14     0 Inf          span.ratio 2.814257e+00    <NA>  500
#> 378      1   1 14     0 Inf          aucinf.obs 2.584803e+01    <NA>  500
#> 379      0   1 14   312 336             auclast 2.257955e+01    <NA>  500
#> 380      0   1 14   312 Inf                cmax 3.385992e+00    <NA>  500
#> 381      0   1 14   312 Inf                tmax 2.000000e+00    <NA>  500
#> 382      0   1 14   312 Inf               tlast 1.680000e+02    <NA>  500
#> 383      0   1 14   312 Inf           clast.obs 9.254052e-04    <NA>  500
#> 384      0   1 14   312 Inf            lambda.z 3.304142e-02    <NA>  500
#> 385      0   1 14   312 Inf           r.squared 9.974514e-01    <NA>  500
#> 386      0   1 14   312 Inf       adj.r.squared 9.968143e-01    <NA>  500
#> 387      0   1 14   312 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 388      0   1 14   312 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 389      0   1 14   312 Inf          clast.pred 8.714471e-04    <NA>  500
#> 390      0   1 14   312 Inf           half.life 2.097813e+01    <NA>  500
#> 391      0   1 14   312 Inf          span.ratio 6.864292e+00    <NA>  500
#> 392      0   1 14   312 Inf          aucinf.obs 2.565471e+01    <NA>  500
#> 393      1   0 15     0  24             auclast 2.703602e+01    <NA>  500
#> 394      1   0 15     0 Inf                cmax 2.495331e+00    <NA>  500
#> 395      1   0 15     0 Inf                tmax 4.000000e+00    <NA>  500
#> 396      1   0 15     0 Inf               tlast 1.680000e+02    <NA>  500
#> 397      1   0 15     0 Inf           clast.obs 2.132697e-03    <NA>  500
#> 398      1   0 15     0 Inf            lambda.z 3.092031e-02    <NA>  500
#> 399      1   0 15     0 Inf           r.squared 9.992950e-01    <NA>  500
#> 400      1   0 15     0 Inf       adj.r.squared 9.990600e-01    <NA>  500
#> 401      1   0 15     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 402      1   0 15     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 403      1   0 15     0 Inf          clast.pred 2.026777e-03    <NA>  500
#> 404      1   0 15     0 Inf           half.life 2.241721e+01    <NA>  500
#> 405      1   0 15     0 Inf          span.ratio 5.353030e+00    <NA>  500
#> 406      1   0 15     0 Inf          aucinf.obs 3.337841e+01    <NA>  500
#> 407      0   0 15   312 336             auclast 2.631952e+01    <NA>  500
#> 408      0   0 15   312 Inf                cmax 3.403290e+00    <NA>  500
#> 409      0   0 15   312 Inf                tmax 3.000000e+00    <NA>  500
#> 410      0   0 15   312 Inf               tlast 1.680000e+02    <NA>  500
#> 411      0   0 15   312 Inf           clast.obs 1.304849e-03    <NA>  500
#> 412      0   0 15   312 Inf            lambda.z 3.330867e-02    <NA>  500
#> 413      0   0 15   312 Inf           r.squared 9.963752e-01    <NA>  500
#> 414      0   0 15   312 Inf       adj.r.squared 9.954690e-01    <NA>  500
#> 415      0   0 15   312 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 416      0   0 15   312 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 417      0   0 15   312 Inf          clast.pred 1.408019e-03    <NA>  500
#> 418      0   0 15   312 Inf           half.life 2.080981e+01    <NA>  500
#> 419      0   0 15   312 Inf          span.ratio 6.919813e+00    <NA>  500
#> 420      0   0 15   312 Inf          aucinf.obs 3.135505e+01    <NA>  500
#> 421      0   1 16     0  24             auclast 2.579005e+01    <NA>  500
#> 422      0   1 16     0 Inf                cmax 4.136596e+00    <NA>  500
#> 423      0   1 16     0 Inf                tmax 3.000000e+00    <NA>  500
#> 424      0   1 16     0 Inf               tlast 1.680000e+02    <NA>  500
#> 425      0   1 16     0 Inf           clast.obs 6.301182e-04    <NA>  500
#> 426      0   1 16     0 Inf            lambda.z 3.968870e-02    <NA>  500
#> 427      0   1 16     0 Inf           r.squared 9.998838e-01    <NA>  500
#> 428      0   1 16     0 Inf       adj.r.squared 9.998451e-01    <NA>  500
#> 429      0   1 16     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 430      0   1 16     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 431      0   1 16     0 Inf          clast.pred 6.159136e-04    <NA>  500
#> 432      0   1 16     0 Inf           half.life 1.746460e+01    <NA>  500
#> 433      0   1 16     0 Inf          span.ratio 6.871043e+00    <NA>  500
#> 434      0   1 16     0 Inf          aucinf.obs 3.094745e+01    <NA>  500
#> 435      1   1 16   312 336             auclast 2.549929e+01    <NA>  500
#> 436      1   1 16   312 Inf                cmax 2.597415e+00    <NA>  500
#> 437      1   1 16   312 Inf                tmax 3.000000e+00    <NA>  500
#> 438      1   1 16   312 Inf               tlast 1.680000e+02    <NA>  500
#> 439      1   1 16   312 Inf           clast.obs 7.585474e-04    <NA>  500
#> 440      1   1 16   312 Inf            lambda.z 3.990737e-02    <NA>  500
#> 441      1   1 16   312 Inf           r.squared 9.994039e-01    <NA>  500
#> 442      1   1 16   312 Inf       adj.r.squared 9.992548e-01    <NA>  500
#> 443      1   1 16   312 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 444      1   1 16   312 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 445      1   1 16   312 Inf          clast.pred 7.096422e-04    <NA>  500
#> 446      1   1 16   312 Inf           half.life 1.736890e+01    <NA>  500
#> 447      1   1 16   312 Inf          span.ratio 8.290679e+00    <NA>  500
#> 448      1   1 16   312 Inf          aucinf.obs 3.116433e+01    <NA>  500
#> 449      1   0 17     0  24             auclast 3.312138e+01    <NA>  500
#> 450      1   0 17     0 Inf                cmax 2.896972e+00    <NA>  500
#> 451      1   0 17     0 Inf                tmax 3.000000e+00    <NA>  500
#> 452      1   0 17     0 Inf               tlast 1.680000e+02    <NA>  500
#> 453      1   0 17     0 Inf           clast.obs 5.010270e-03    <NA>  500
#> 454      1   0 17     0 Inf            lambda.z 2.474493e-02    <NA>  500
#> 455      1   0 17     0 Inf           r.squared 9.956135e-01    <NA>  500
#> 456      1   0 17     0 Inf       adj.r.squared 9.941513e-01    <NA>  500
#> 457      1   0 17     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 458      1   0 17     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 459      1   0 17     0 Inf          clast.pred 5.012427e-03    <NA>  500
#> 460      1   0 17     0 Inf           half.life 2.801168e+01    <NA>  500
#> 461      1   0 17     0 Inf          span.ratio 4.283927e+00    <NA>  500
#> 462      1   0 17     0 Inf          aucinf.obs 4.200587e+01    <NA>  500
#> 463      0   0 17   312 336             auclast 3.313007e+01    <NA>  500
#> 464      0   0 17   312 Inf                cmax 3.989345e+00    <NA>  500
#> 465      0   0 17   312 Inf                tmax 1.500000e+00    <NA>  500
#> 466      0   0 17   312 Inf               tlast 1.680000e+02    <NA>  500
#> 467      0   0 17   312 Inf           clast.obs 5.272342e-03    <NA>  500
#> 468      0   0 17   312 Inf            lambda.z 2.311936e-02    <NA>  500
#> 469      0   0 17   312 Inf           r.squared 9.985419e-01    <NA>  500
#> 470      0   0 17   312 Inf       adj.r.squared 9.980558e-01    <NA>  500
#> 471      0   0 17   312 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 472      0   0 17   312 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 473      0   0 17   312 Inf          clast.pred 5.224492e-03    <NA>  500
#> 474      0   0 17   312 Inf           half.life 2.998124e+01    <NA>  500
#> 475      0   0 17   312 Inf          span.ratio 4.002503e+00    <NA>  500
#> 476      0   0 17   312 Inf          aucinf.obs 4.104374e+01    <NA>  500
#> 477      1   1 18     0  24             auclast 2.774108e+01    <NA>  500
#> 478      1   1 18     0 Inf                cmax 3.490646e+00    <NA>  500
#> 479      1   1 18     0 Inf                tmax 3.000000e+00    <NA>  500
#> 480      1   1 18     0 Inf               tlast 1.680000e+02    <NA>  500
#> 481      1   1 18     0 Inf           clast.obs 8.076641e-04    <NA>  500
#> 482      1   1 18     0 Inf            lambda.z 3.561665e-02    <NA>  500
#> 483      1   1 18     0 Inf           r.squared 9.999991e-01    <NA>  500
#> 484      1   1 18     0 Inf       adj.r.squared 9.999981e-01    <NA>  500
#> 485      1   1 18     0 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 486      1   1 18     0 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 487      1   1 18     0 Inf          clast.pred 8.068921e-04    <NA>  500
#> 488      1   1 18     0 Inf           half.life 1.946132e+01    <NA>  500
#> 489      1   1 18     0 Inf          span.ratio 3.699646e+00    <NA>  500
#> 490      1   1 18     0 Inf          aucinf.obs 3.241357e+01    <NA>  500
#> 491      0   1 18   312 336             auclast 2.559760e+01    <NA>  500
#> 492      0   1 18   312 Inf                cmax 3.736891e+00    <NA>  500
#> 493      0   1 18   312 Inf                tmax 2.000000e+00    <NA>  500
#> 494      0   1 18   312 Inf               tlast 1.680000e+02    <NA>  500
#> 495      0   1 18   312 Inf           clast.obs 7.769251e-04    <NA>  500
#> 496      0   1 18   312 Inf            lambda.z 3.656346e-02    <NA>  500
#> 497      0   1 18   312 Inf           r.squared 9.999991e-01    <NA>  500
#> 498      0   1 18   312 Inf       adj.r.squared 9.999983e-01    <NA>  500
#> 499      0   1 18   312 Inf lambda.z.time.first 9.600000e+01    <NA>  500
#> 500      0   1 18   312 Inf   lambda.z.n.points 3.000000e+00    <NA>  500
#> 501      0   1 18   312 Inf          clast.pred 7.761936e-04    <NA>  500
#> 502      0   1 18   312 Inf           half.life 1.895737e+01    <NA>  500
#> 503      0   1 18   312 Inf          span.ratio 3.797994e+00    <NA>  500
#> 504      0   1 18   312 Inf          aucinf.obs 3.045015e+01    <NA>  500
#> 505      0   1 19     0  24             auclast 2.079130e+01    <NA>  500
#> 506      0   1 19     0 Inf                cmax 3.904765e+00    <NA>  500
#> 507      0   1 19     0 Inf                tmax 2.000000e+00    <NA>  500
#> 508      0   1 19     0 Inf               tlast 1.680000e+02    <NA>  500
#> 509      0   1 19     0 Inf           clast.obs 1.327456e-04    <NA>  500
#> 510      0   1 19     0 Inf            lambda.z 4.779997e-02    <NA>  500
#> 511      0   1 19     0 Inf           r.squared 9.993155e-01    <NA>  500
#> 512      0   1 19     0 Inf       adj.r.squared 9.991444e-01    <NA>  500
#> 513      0   1 19     0 Inf lambda.z.time.first 2.400000e+01    <NA>  500
#> 514      0   1 19     0 Inf   lambda.z.n.points 6.000000e+00    <NA>  500
#> 515      0   1 19     0 Inf          clast.pred 1.255944e-04    <NA>  500
#> 516      0   1 19     0 Inf           half.life 1.450100e+01    <NA>  500
#> 517      0   1 19     0 Inf          span.ratio 9.930351e+00    <NA>  500
#> 518      0   1 19     0 Inf          aucinf.obs 2.338106e+01    <NA>  500
#> 519      1   1 19   312 336             auclast 2.062020e+01    <NA>  500
#> 520      1   1 19   312 Inf                cmax 2.727210e+00    <NA>  500
#> 521      1   1 19   312 Inf                tmax 3.000000e+00    <NA>  500
#> 522      1   1 19   312 Inf               tlast 1.680000e+02    <NA>  500
#> 523      1   1 19   312 Inf           clast.obs 1.401087e-04    <NA>  500
#> 524      1   1 19   312 Inf            lambda.z 4.659617e-02    <NA>  500
#> 525      1   1 19   312 Inf           r.squared 9.995674e-01    <NA>  500
#> 526      1   1 19   312 Inf       adj.r.squared 9.993510e-01    <NA>  500
#> 527      1   1 19   312 Inf lambda.z.time.first 7.200000e+01    <NA>  500
#> 528      1   1 19   312 Inf   lambda.z.n.points 4.000000e+00    <NA>  500
#> 529      1   1 19   312 Inf          clast.pred 1.343427e-04    <NA>  500
#> 530      1   1 19   312 Inf           half.life 1.487562e+01    <NA>  500
#> 531      1   1 19   312 Inf          span.ratio 6.453510e+00    <NA>  500
#> 532      1   1 19   312 Inf          aucinf.obs 2.347797e+01    <NA>  500
#> 533      0   0 20     0  24             auclast 4.026034e+01    <NA>  500
#> 534      0   0 20     0 Inf                cmax 4.909493e+00    <NA>  500
#> 535      0   0 20     0 Inf                tmax 2.000000e+00    <NA>  500
#> 536      0   0 20     0 Inf               tlast 1.680000e+02    <NA>  500
#> 537      0   0 20     0 Inf           clast.obs 8.320385e-03    <NA>  500
#> 538      0   0 20     0 Inf            lambda.z 2.631276e-02    <NA>  500
#> 539      0   0 20     0 Inf           r.squared 9.998031e-01    <NA>  500
#> 540      0   0 20     0 Inf       adj.r.squared 9.997375e-01    <NA>  500
#> 541      0   0 20     0 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 542      0   0 20     0 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 543      0   0 20     0 Inf          clast.pred 8.405289e-03    <NA>  500
#> 544      0   0 20     0 Inf           half.life 2.634262e+01    <NA>  500
#> 545      0   0 20     0 Inf          span.ratio 4.555355e+00    <NA>  500
#> 546      0   0 20     0 Inf          aucinf.obs 5.535127e+01    <NA>  500
#> 547      1   0 20   312 336             auclast 3.686500e+01    <NA>  500
#> 548      1   0 20   312 Inf                cmax 2.751239e+00    <NA>  500
#> 549      1   0 20   312 Inf                tmax 6.000000e+00    <NA>  500
#> 550      1   0 20   312 Inf               tlast 1.680000e+02    <NA>  500
#> 551      1   0 20   312 Inf           clast.obs 8.646309e-03    <NA>  500
#> 552      1   0 20   312 Inf            lambda.z 2.776800e-02    <NA>  500
#> 553      1   0 20   312 Inf           r.squared 9.909760e-01    <NA>  500
#> 554      1   0 20   312 Inf       adj.r.squared 9.879680e-01    <NA>  500
#> 555      1   0 20   312 Inf lambda.z.time.first 4.800000e+01    <NA>  500
#> 556      1   0 20   312 Inf   lambda.z.n.points 5.000000e+00    <NA>  500
#> 557      1   0 20   312 Inf          clast.pred 7.981696e-03    <NA>  500
#> 558      1   0 20   312 Inf           half.life 2.496209e+01    <NA>  500
#> 559      1   0 20   312 Inf          span.ratio 4.807290e+00    <NA>  500
#> 560      1   0 20   312 Inf          aucinf.obs 5.416901e+01    <NA>  500