NIF file debugger

This shiny app helps identifying suspicious data in the NIF data set. This may be due to missing data in the underlying SDTM data set or bad imputations during the generation of the NIF data set.

Usage

nif_debugger(
  nif_data,
  sdtm_data,
  analyte = NULL,
  extrt = NULL,
  pctestcd = NULL,
  usubjid = NULL
)

Arguments

nif_data

A NIF data set.

sdtm_data

The underlying SDTM data set.

analyte

The analyte as character. If NULL (default), the most likely analyte will be selected.

extrt

The EXTRT field.

pctestcd

The PCTESTCD field.

usubjid

The subject USUBJID to filter for. If NULL (default), all subjects are shown.

Value

Nothing