Plot nif object.

Plot nif object.

Usage

plot_old(
  x,
  analyte = NULL,
  dose = NULL,
  log = FALSE,
  time = "TAFD",
  group = NULL,
  min_time = NULL,
  max_time = NULL,
  points = FALSE,
  lines = TRUE,
  admin = NULL,
  cfb = FALSE,
  summary_function = median,
  mean = FALSE,
  title = "",
  caption = "",
  integrate_predose = TRUE,
  legend = TRUE,
  show_n = FALSE,
  shading = TRUE,
  silent = deprecated(),
  dose_norm = FALSE,
  .by = "DOSE",
  ...
)

Arguments

x

A nif object.

analyte

The analyte as character.

dose

The dose as numeric.

log

Logarithmic y scale

time

The time field to use, as character. Can be 'TIME', 'NTIME', 'TAFD' or 'TAD'.

group

The grouping variable as character.

min_time

The minimal time as numeric.

max_time

The maximal time as numeric.

points

Show data points as logical.

lines

Show lines as logical.

admin

Show vertical lines for administrations, as logical.

cfb

Show change from baseline, as logical.

summary_function

The summarizing function to apply to multiple baseline values.

mean

Show a mean plot, as logical.

title

The plot title as character.

caption

The plot caption line as character.

integrate_predose

Complete 'DOSE' field for predose values.

legend

Show legend, as logical.

show_n

Show sample size in mean plot, as logical. Does currently not implement grouping!

shading

Show ribbons when mean = TRUE, as logical.

silent

Dummy option for compatibility, set the global option nif_option() with silent = TRUE to suppress messages.

dose_norm

Dose-normalized exposure, as logical.

.by

Facetting variable, as character.

...

Further graphical parameters for geom_point().

Value

A ggplot object.

Details

A watermark can be added by setting the global option 'watermark', e.g., nif_option(watermark = "text").

See also