Load SDTM data from sas7bdat files

read_sdtm_sas() has been deprecated in favor of read_sdtm().

Usage

read_sdtm_sas(data_path, ...)

Arguments

data_path

Path to the required SDTM files in .sas7bdat format.

...

Further optional parameters may specify the individual SDTM domains to be loaded (lowercase, no file extensions). If no further parameters are given, the standard set of dm, ex, pc and vs are loaded.

Value

A named list of the relevant SDTM domains as data.frames

Details

This function loads relevant SDTM domain files required for the generation of NIF files from a folder. It assumes that DM, EX, PC and VS are available at this location and will throw an error if not.

See also