Synthesize baseline serum creatinine for fictional subjects

Synthesize baseline serum creatinine for fictional subjects

Usage

synthesize_crea(dm, crea_method = crea_mdrd)

Arguments

dm

The DM domain as data frame.

crea_method

The creatinine calculation function as function reference. Can currently be crea_mdrd or crea_raynaud.

Value

A DM domain with additional fields as data frame.

Details

This function makes use of empirical eGFR data from a study in non-diseased Caucasian subjects as published by Wetzels et al., Table 1. In this study, eGFR were calculated by the Modification of Diet in Renal Disease (MDRD) method.

The tabulated data is modeled using a generalized linear model based on age and sex. The model is used to predict the target eGFR (target_egfr) for the subjects in DM. An actual eGFR (egfr) is then generated for each subjects as a random sample from a normal distribution with the target eGFR as mean and a standard deviation of 13 (i.e., the mean SD in the empirical data).

A corresponding baseline creatinine concentration is then calculated from the actual eGFR using the method specified in crea_method.