Synthesize a fictional DM domain

Currently geared toward a food effect study, but can be adapted easily to generate other study types.

Usage

synthesize_dm(
  studyid = "2023001",
  nsubs = 10,
  nsites = 5,
  duration = 7,
  female_fraction = 0.5,
  min_age = 18,
  max_age = 55
)

Arguments

studyid

The study ID.

nsubs

The number of subjects.

nsites

The number of clinical sites.

duration

The duration of the study in days.

female_fraction

The fraction of female subjects (between 0 and 1).

min_age

The minimum age.

max_age

The maximum age.

Value

The DM data as data frame.