Attach a parent-metabolite mapping to a SDTM object.

In case multiple analytes are measured for a specific administered drug, some functions need that information to correlate plasma concentrations with administrations. 'add_metabolite_mapping()' is used to attach this information to a SDTM object.

Usage

add_metabolite_mapping(obj, pctestcd_parent, pctestcd_metabolite)

Arguments

obj

The SDTM object.

pctestcd_parent

The PCTESTCD of the parent compound.

pctestcd_metabolite

The PCTESTCD of the metabolite.

Value

The SDTM object.

See also

Examples

add_metabolite_mapping(examplinib_fe, "RS2023", "RS2023487A")
#> -------- SDTM data set summary -------- 
#> Study 2023000400 with 20 subjects providing PC data.
#> 
#> Subjects per domain:
#>  DOMAIN   N    
#>  dm       27   
#>  vs       27   
#>  ex       20   
#>  pc       20   
#>  lb       27   
#>  pp       20   
#> 
#> Arms (DM):
#>  ACTARM           ACTARMCD   
#>  Fed - Fasted     BA         
#>  Screen Failure   SCRNFAIL   
#>  Fasted - Fed     AB         
#> 
#> Treatments (EX):
#>  EXTRT        
#>  EXAMPLINIB    
#> 
#> Specimens (PC):
#>  none 
#> 
#> Analytes (PC):
#>  PCTEST       PCTESTCD     
#>  RS2023       RS2023       
#>  RS2023487A   RS2023487A    
#> 
#> Treatment-to-analyte mappings:
#>  EXTRT        PCTESTCD   ANALYTE   
#>  EXAMPLINIB   RS2023     RS2023     
#> 
#> Parent-to-metabolite mappings:
#>  PCTESTCD_parent   PCTESTCD_metab   
#>  RS2023            RS2023487A        
#> 
#> Time mappings:
#>  none 
#>