Append observation events

Observations can be pharmacokinetic observations (i.e., from the PC domain), or any other class of observation from any other SDTM domain. The 'testcd' specifies the value of the respective TESTCD field (e.g., 'PCTESTCD', 'VSTESTCD' or 'LBTESTCD') that defines the observation. Observation events can be attached to an administered drug by specifying the 'parent' field. This is required for, e.g., the time-after-dose ('TAD') and time-after-first-dose ('TAFD') time calculation.

Usage

add_observation(
  nif,
  sdtm,
  domain,
  testcd,
  analyte = NULL,
  parent = NULL,
  cmt = NULL,
  subject_filter = "!ACTARMCD %in% c('SCRNFAIL', 'NOTTRT')",
  observation_filter = "TRUE",
  TESTCD_field = NULL,
  DTC_field = NULL,
  DV_field = NULL,
  factor = 1,
  NTIME_lookup = NULL,
  keep = NULL,
  silent = deprecated()
)

Arguments

nif

A nif object.

sdtm

A sdtm object. Needs at least the 'DM' and 'VS' domains, and the domain the observations come from.

domain

The domain as character.

testcd

The observation variable, as character.

analyte

The name for the analyte. Defaults to the 'testcd', if NULL.

parent

The name of the parent analyte for the observation as character.

cmt

The compartment for the observation as numeric.

subject_filter

The filtering to apply to the DM domain.

observation_filter

The filtering to apply to the observation source data.

TESTCD_field

The xxTESTCD field. Defaults to the two-character domain name followed by 'TESTCD', if NULL.

DTC_field

The field to use as the date-time code for the observation. Defaults to the two-character domain name followed by 'DTC', if NULL.

DV_field

the field to use as the dependent variable. Defaults to the two-character domain name followed by 'STRESN', if NULL.

factor

Multiplier for the DV field, as numeric.

NTIME_lookup

A data frame with two columns, a column that defines the custom nominal time information in the target domain (e.g., 'PCELTM'), and 'NTIME'. This data frame is left_joined into the observation data frame to provide the NTIME field.

keep

Columns to keep, as character.

silent

Dummy option for compatibility, set the global option nif_option() with silent = TRUE to suppress messages.

Value

A nif object.

Details

Observations can be further specified with the 'observation_filter' term. The filter term can refer to any field of the respective SDTM domain.

A PK/PD model compartment can be specified with 'cmt' or will be automatically assigned if cmt = NULL.

See also

Examples

add_observation(examplinib_fe_nif, examplinib_fe, "pc", "RS2023487A",
  parent = "RS2023")
#> Compartment for RS2023487A was not specified and has been set to 3
#> ----- NONMEM input file (NIF) object -----
#> RS2023 data from one study
#> 1360 observations from 20 subjects
#> Males: 17, females: 3 (15%)
#> 
#> Columns:
#> REF, ID, STUDYID, USUBJID, AGE, SEX, RACE, HEIGHT, WEIGHT, BMI, DTC, TIME, NTIME, TAFD, TAD, PCELTM, EVID, AMT, ANALYTE, CMT, PARENT, TRTDY, METABOLITE, DOSE, DV, MDV, ACTARMCD, IMPUTATION, PERIOD, FASTED, TREATMENT 
#> 
#> NIF data (selected columns):
#> ID   NTIME   TIME   TAD   ANALYTE      EVID   CMT   AMT   DOSE   DV         
#> 1    0       0      0     RS2023       1      1     500   500    NA         
#> 1    0       0      0     RS2023       0      2     0     500    0          
#> 1    0       0      0     RS2023487A   0      3     0     500    0          
#> 1    0.5     0.5    0.5   RS2023       0      2     0     500    249.493    
#> 1    0.5     0.5    0.5   RS2023487A   0      3     0     500    130.867    
#> 1    1       1      1     RS2023       0      2     0     500    858.22     
#> 1    1       1      1     RS2023487A   0      3     0     500    845.214    
#> 1    1.5     1.5    1.5   RS2023       0      2     0     500    1767.312   
#> 1    1.5     1.5    1.5   RS2023487A   0      3     0     500    2321.885   
#> 1    2       2      2     RS2023       0      2     0     500    2176.502   
#> 1390 more rows