Make a NIF object from SDTM-formatted data

This function creates a basic NONMEM input file (NIF) data set from SDTM-formatted clinical study data following the conventions summarized in Bauer, CPT Pharmacometrics Syst. Pharmacol. (2019). For a more in-depth tutorial, see vignette("nif-vignette").

Usage

make_nif(
  sdtm_data,
  spec = NULL,
  silent = FALSE,
  truncate_to_last_observation = TRUE,
  truncate_to_last_individual_obs = FALSE,
  use_pctptnum = FALSE,
  analyte_cmt_mapping = NULL
)

Arguments

sdtm_data

A sdtm object, i.e., essentially a list of SDTM domains as data tables. Typically, the SDTM data are loaded using read_sdtm_sas() or read_sdtm_xpt(). As a minimum, the following SDTM domains are needed: DM, VS, PC and EX.

spec

The sample specimen for the PC data as string (e.g., "BLOOD", "PLASMA", "URINE", "FECES"). When spec is NULL (default), the most likely specimen is selected.

silent

Switch to disable message output.

truncate_to_last_observation

Boolean to indicate whether the data set should be truncated to the last observation. In this case, administrations after the last observation time point will deleted. The default is 'TRUE'.

truncate_to_last_individual_obs

Boolean to indicate whether observations should be truncted to the last individual observation.

use_pctptnum

Boolean to indicate whether to derive nominal time ('NTIME') from 'PCTPTNUM'.

analyte_cmt_mapping

The analyte-compartment association as data frame with the columns 'ANALYTE' and 'CMT'.

Value

A NIF object.

Imputations

Subjects with administration but no observations for the respective analyte are deleted from the data set. For further imputations, see vignette("nif-imputations").

Output fields

ID Subject identification number

TIME Recorded time of administration or observation events in hours relative to the first individual event.

AMT Dose administered for dosing record, or zero for observations.

DOSE Dose in mg for administrations and post-dose observations.

DV The dependent variable, i.e., observed concentration, or zero for administration records, in mg/l.

LNDV The natural Log of DV.

RATE Rate of infusion of drug or zero if drug is given as a bolus.

MDV One for missing DV, else zero.

EVID Event ID: 0 for observations, 1 for administrations.

CMT Pharmacokinetic compartment. Will be set to 1 for administrations and 2 for observations. Should be changed afterwards, if needed.

DTC The date-time of the data record.

FIRSTDTC Date and time of first event per subject. This field is used internally for the calculation of TIME. Although it is not needed for NONMEM analysis, it is provided for subsequent NIF file building steps, e.g., addition of further time-dependent endpoints.

FIRSTADMINDTC The date-time of the first administration of the respective parent drug for the respective subject.

FIRSTTRTDTC The date-time of the first administration of any parent drug for the respective subject.

ANALYTE The analyte or drug in the data record.

TRTDY The treatment day, i.e., the relative day after the first treatment for the respective subject.

See also