Automatically generate a nif object with pharmacokinetic observations

Usage

nif_auto(sdtm, bl_creat = TRUE, bl_odwg = TRUE, keep = NULL)

Arguments

sdtm

The source SDTM object.

bl_creat

Include baseline creatinine, creatinine clearance and renal function class fields.

bl_odwg

Include baseline ODWG hepatic function class.

keep

Columns to keep, as character.

Value

A nif object.

Details

Automatically generate a nif object from SDTM data. All treatments and analytes must be defined in the treatment_mapping and metabolite_mapping objects that are attached to the sdtm object.

See also

Examples

nif_auto(examplinib_sad)
#> Adding treatment(s): EXAMPLINIB
#> Adding PC observations(s): RS2023
#> Compartment for RS2023 was not specified and has been set to 2
#> Adding baseline renal function
#> Adding baseline hepatic function
#> ----- NONMEM input file (NIF) object -----
#> RS2023 data from one study
#> 816 observations from 48 subjects
#> Males: 48, females: 0 (0%)
#> 
#> Columns:
#> REF, ID, STUDYID, USUBJID, AGE, SEX, RACE, HEIGHT, WEIGHT, BMI, DTC, TIME, NTIME, TAFD, TAD, PCELTM, EVID, AMT, ANALYTE, CMT, PARENT, TRTDY, METABOLITE, DOSE, DV, MDV, ACTARMCD, IMPUTATION, BL_CREAT, BL_CRCL, BL_RENAL, BL_BILI_X_ULN, BL_AST_X_ULN, BL_ODWG 
#> 
#> NIF data (selected columns):
#> ID   NTIME   TIME   TAD   ANALYTE   EVID   CMT   AMT   DOSE   DV       
#> 1    0       0      0     RS2023    1      1     5     5      NA       
#> 1    0       0      0     RS2023    0      2     0     5      0        
#> 1    0.5     0.5    0.5   RS2023    0      2     0     5      2.447    
#> 1    1       1      1     RS2023    0      2     0     5      7.245    
#> 1    1.5     1.5    1.5   RS2023    0      2     0     5      15.748   
#> 1    2       2      2     RS2023    0      2     0     5      20.923   
#> 1    3       3      3     RS2023    0      2     0     5      14.767   
#> 1    4       4      4     RS2023    0      2     0     5      13.274   
#> 1    6       6      6     RS2023    0      2     0     5      5.484    
#> 1    8       8      8     RS2023    0      2     0     5      2.124    
#> 854 more rows